requirements.txt
amqp==5.1.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
crispy-bootstrap5==0.6
cryptography==36.0.1
Deprecated==1.2.13
distro==1.7.0
dj-database-url==0.5.0
Django==3.2.13
django-crispy-forms==1.14.0
django-timezone-field==4.2.3
django-celery-beat==2.2.1
djangorestframework==3.12.4
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fonttools==4.26.2
gunicorn==20.1.0
idna==3.3
jsonschema==3.2.0
kombu==5.2.4
mysql-connector-python==8.0.27
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
paramiko==2.10.1
Pillow==8.3.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-gettext==4.0
pytz==2022.1
PyYAML==5.4.1
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
texttable==1.6.4
ttf2web==0.9.3
tweepy==4.8.0
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
whitenoise==6.0.0
wrapt==1.14.0
